+++ src/interfaces.d.ts
